+++ packages/node/CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased - 2023-06-20
+## [Unreleased]
 ### Added
-- Add flag `query-address-limit` for dictionary queries
+- Add flag `query-address-limit` for dictionary queries (#111)
 
 ## [2.6.1] - 2023-06-19
 ### Added